refactor: add custom toggle button to new hierarchy
