+++ src/gui/components/buttons/v2/SMPAbstractToggleButton.java
@@ -1,0 +1,35 @@
+package gui.components.buttons.v2;
+
+import java.util.function.Supplier;
+
+import javafx.beans.property.BooleanProperty;
+import javafx.scene.control.Skin;
+import javafx.scene.control.ToggleButton;
+
+public abstract class SMPAbstractToggleButton<B extends ToggleButton> extends SMPAbstractButton<B> {
+
+	protected SMPAbstractToggleButton(Supplier<B> fnCreateInnerBtn) {
+		super(fnCreateInnerBtn);
+	}
+	
+	@Override
+	public BooleanProperty selectedProperty() {
+		return innerBtn.selectedProperty();
+	}
+	
+	@Override
+    public boolean isSelected() {
+        return innerBtn.isSelected();
+    }
+	
+	@Override
+	public void setSelected(boolean value) {
+        innerBtn.setSelected(value);
+    }
+	
+	@Override
+	protected Skin<?> createDefaultSkin() {
+		return new SMPToggleButtonSkin<>(this);
+	}
+	
+}
